# backend/openapi.yaml
openapi: 3.0.3
info:
  title: progctf — The Voyage API
  version: 0.1.0
  description: |
    Pirate-themed CTF backend (Bosun). Player REST API + admin API.
    JWT auth via httpOnly cookie (`progctf_session` for player, `progctf_admin` for admin)
    or `Authorization: Bearer <token>` header.
servers:
  - url: http://localhost:4000
    description: Local dev
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: progctf_session
    adminCookie:
      type: apiKey
      in: cookie
      name: progctf_admin
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [ok, code, message]
      properties:
        ok: { type: boolean, example: false }
        code:
          type: string
          enum:
            - ERR_BAD_INPUT
            - ERR_BAD_FORMAT
            - ERR_BAD_CREDENTIALS
            - ERR_PIRATE_EXISTS
            - ERR_BANNED
            - ERR_NO_AUTH
            - ERR_BAD_AUTH
            - ERR_NOT_ADMIN
            - ERR_NO_REFRESH
            - ERR_BAD_REFRESH
            - ERR_NO_CREW
            - ERR_ALREADY_ON_CREW
            - ERR_CREW_NAME_TAKEN
            - ERR_CREW_BANNED
            - ERR_NO_ISLAND
            - ERR_NOT_PUBLISHED
            - ERR_NO_WHISPER
            - ERR_BAD_WHISPER
            - ERR_RATE_LIMITED
            - ERR_ALREADY_PLUNDERED
            - ERR_WRONG_TREASURE
            - ERR_VOYAGE_FROZEN
            - ERR_SLUG_TAKEN
            - ERR_WHISPER_EXISTS
            - ERR_NO_PIRATE
            - ERR_NO_UPDATES
            - ERR_SANDBOX_DOWN
        message: { type: string, example: "That ain't the right Treasure, sailor." }
        retryAfter: { type: integer }
...
    IslandList:
      type: object
      properties:
        slug: { type: string }
        title: { type: string }
        category: { type: string }
        difficulty: { type: string, enum: [port, open_sea, cursed_depths] }
        base_points: { type: integer }
        current_points: { type: integer }
        first_blood_crew: { type: string, nullable: true }
        solves: { type: integer }
        solved: { type: boolean }
    IslandDetail:
      allOf:
        - $ref: '#/components/schemas/IslandList'
        - type: object
          properties:
            description_md: { type: string }
            files_url: { type: string, nullable: true }
            sandbox_image: { type: string, nullable: true }
            solved_by_crew: { type: boolean }
            whispers:
              type: array
              items:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  ordinal: { type: integer }
                  cost_points: { type: integer }
                  revealed: { type: boolean }
                  body_md: { type: string, nullable: true }
    SubmissionAccepted:
      type: object
      properties:
        ok: { type: boolean, example: true }
        awarded: { type: integer }
        baseAward: { type: integer }
        firstBlood: { type: integer }
        currentPoints: { type: integer }
        islandSlug: { type: string }
        islandTitle: { type: string }
        message: { type: string }
...
paths:
...
  /api/islands/:
    get:
      tags: [islands]
      summary: |
        List Islands. By default returns only `status='published'` (sanitized — no flag_hash).
        Admins may pass `?include_unpublished=1` to also see drafts and archived. Non-admins
        passing the param get 403 (we do NOT silently ignore — frontend needs to know).
      security: [{ sessionCookie: [] }, { bearer: [] }]
      parameters:
        - in: query
          name: include_unpublished
          required: false
          description: |
            "1" or "true" — admin-only. Includes drafts and archived. 403 for non-admins.
          schema:
            type: string
            enum: ['1', 'true']
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  islands:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/IslandList'
                        - type: object
                          properties:
                            status: { type: string, enum: [draft, published, archived] }
        '403':
          description: non-admin tried to use include_unpublished
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /api/islands/{slug}:
    get:
      tags: [islands]
      summary: Island detail
      security: [{ sessionCookie: [] }, { bearer: [] }]
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  island: { $ref: '#/components/schemas/IslandDetail' }
        '404': { description: not found / not published }
  /api/islands/{slug}/submit:
    post:
      tags: [islands]
      summary: Submit a Treasure (rate-limited 5/min/Crew/Island)
      security: [{ sessionCookie: [] }, { bearer: [] }]
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [flag]
              properties:
                flag: { type: string, example: 'progctf{set_sail}' }
      responses:
        '200':
          description: Either correct (SubmissionAccepted) OR wrong-treasure error (still 200)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/SubmissionAccepted'
                  - $ref: '#/components/schemas/ErrorResponse'
        '400': { description: malformed }
        '403': { description: no crew }
        '404': { description: no such island / not published }
        '409': { description: already plundered }
        '423': { description: voyage frozen }
        '429':
          description: rate limited
          headers:
            Retry-After:
              schema: { type: integer }
  /api/islands/{slug}/shell:
    get:
      tags: [islands]
      summary: |
        WebSocket — per-Island terminal stub. The HTTP GET shown here is the WS
        upgrade request; clients connect via `ws://.../api/islands/{slug}/shell`
        with the player JWT cookie or `Authorization: Bearer <token>`.

        Phase 1c behaviour: server emits a JSON `banner` frame on connect, then
        for each inbound message replies with a JSON `stdout` frame containing
        `"⚓ shell stub for {slug} — sandbox controller hookup is Phase 2"`.

        Phase 2 will wire each socket to the per-Crew sandbox container managed
        by Shipwright + Powder Monkey.
      security: [{ sessionCookie: [] }, { bearer: [] }]
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        '101':
          description: WebSocket upgrade accepted
        '401':
          description: missing or invalid auth — upgrade refused
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  /api/islands/{slug}/whisper/{n}:
    post:
      tags: [islands]
      summary: Reveal Whisper n (deducts points on next solve)
      security: [{ sessionCookie: [] }, { bearer: [] }]
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
        - in: path
          name: n
          required: true
          schema: { type: integer, minimum: 1, maximum: 3 }
      responses:
        '200': { description: ok }
...
  /admin/islands/:
    post:
      tags: [admin]
      summary: Create Island
      security: [{ adminCookie: [] }, { bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug, title, category, difficulty, description_md, base_points, flag]
              properties:
                slug: { type: string }
                title: { type: string }
                category: { type: string, enum: [cursed_ports, cipher_cove, shipwrights_forge, lighthouse, crows_nest, hidden_cargo, keymaster] }
                difficulty: { type: string, enum: [port, open_sea, cursed_depths] }
                description_md: { type: string }
                base_points: { type: integer, minimum: 50, maximum: 900 }
                flag: { type: string, pattern: '^progctf\{[a-z0-9_]+\}$' }
                files_url: { type: string, nullable: true }
                sandbox_image: { type: string, nullable: true }
                status: { type: string, enum: [draft, published, archived], default: draft }
      responses:
        '201': { description: created }
        '409': { description: slug taken }
  /admin/islands/{id}:
    patch:
      tags: [admin]
      summary: Edit Island (any subset of fields)
      security: [{ adminCookie: [] }, { bearer: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: updated }
        '404': { description: not found }
    delete:
      tags: [admin]
      summary: Delete Island
      security: [{ adminCookie: [] }, { bearer: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: scuttled }
...